renamed sign-in and deleted unused sign up page now that were using amplify Authenticator
